Added new behaviors when a user is part of a house
default.py :
- add_house( ): Added creating and joining variables to check whether
user wants to create or join a house. Uses form when creating. joining
not implemented yet.
- people ( ): Added house_list that takes all houses that user is a
member of.

table.py : Took out the 'text' parameter on title so that title form is
smaller.

add_house.html : Added Joining and Creating behaviors

house.html and people.html : switched most of the placeholder info with
automated info that will aggregate once user is part of a house.

user.html : Made the Login page cleaner and centered it to fit with the
look of our app.